requirements.txt
flask
firebase
firebase-admin
Pyrebase4